refactor(editor): studio thumbnail panel 분리
