+++ front/src/routes/Admin/EditorStudioPage.tsx
@@ -47,6 +47,10 @@ import { useEditorStudioRouting } from "./useEditorStudioRouting"
 import { useEditorStudioThumbnailControls } from "./useEditorStudioThumbnailControls"
 import { useEditorStudioThumbnailPreview } from "./useEditorStudioThumbnailPreview"
 import {
+  EditorStudioThumbnailEditorPanel,
+  EditorStudioThumbnailMetaPanel,
+} from "./EditorStudioThumbnailPanels"
+import {
   isServerTempDraftPost,
   TEMP_DRAFT_BODY_PLACEHOLDER,
 } from "./editorTempDraft"
@@ -2286,132 +2290,36 @@ export const EditorStudioPage: NextPage<AdminPageProps> = ({ initialMember }) =>
       zoom: DEFAULT_THUMBNAIL_ZOOM,
     })
   }, [commitPreviewThumbTransform, previewThumbTransformRef])
-  const thumbnailEditorPanel = useMemo(() => (
-    <PreviewEditorSection>
-      <PreviewEditorSectionHeader>
-        <strong>썸네일 위치 조정</strong>
-      </PreviewEditorSectionHeader>
-      <PreviewThumbFrame
-        ref={previewThumbFrameRef}
-        data-draggable={safePreviewThumbnail && !isPreviewThumbnailError}
-        data-dragging={isPreviewThumbDragging}
-        onPointerDown={handlePreviewThumbPointerDown}
-        onPointerMove={handlePreviewThumbPointerMove}
-        onPointerUp={finalizePreviewThumbPointer}
-        onPointerCancel={finalizePreviewThumbPointer}
-      >
-        {safePreviewThumbnail && !isPreviewThumbnailError ? (
-          // eslint-disable-next-line @next/next/no-img-element
-          <img
-            src={safePreviewThumbnail}
-            alt="포스트 미리보기 썸네일"
-            style={{
-              width: "var(--preview-thumb-width)",
-              height: "var(--preview-thumb-height)",
-              left: "var(--preview-thumb-left)",
-              top: "var(--preview-thumb-top)",
-              maxWidth: "none",
-              transform: "translateZ(0)",
-            }}
-            onError={() => setIsPreviewThumbnailError(true)}
-          />
-        ) : (
-          <div className="placeholder">
-            <em>썸네일 없음</em>
-          </div>
-        )}
-      </PreviewThumbFrame>
-      {safePreviewThumbnail && !isPreviewThumbnailError ? (
-        <ZoomControlRow>
-          <FieldLabel htmlFor="post-thumbnail-zoom-modal">썸네일 배율</FieldLabel>
-          <ZoomRangeInput
-            id="post-thumbnail-zoom-modal"
-            type="range"
-            min={1}
-            max={2.5}
-            step={0.01}
-            value={postThumbnailZoom}
-            onChange={(e) => handleThumbnailZoomModalChange(Number(e.target.value))}
-          />
-          <ZoomControlMeta>
-            <ZoomValue>{postThumbnailZoom.toFixed(2)}x</ZoomValue>
-            <Button type="button" onClick={resetThumbnailZoomInModal}>
-              배율 초기화
-            </Button>
-          </ZoomControlMeta>
-        </ZoomControlRow>
-      ) : null}
-    </PreviewEditorSection>
-  ), [
-    finalizePreviewThumbPointer,
-    handlePreviewThumbPointerDown,
-    handlePreviewThumbPointerMove,
-    handleThumbnailZoomModalChange,
-    isPreviewThumbDragging,
-    isPreviewThumbnailError,
-    postThumbnailZoom,
-    previewThumbFrameRef,
-    resetThumbnailZoomInModal,
-    safePreviewThumbnail,
-    setIsPreviewThumbnailError,
-  ])
-  const previewMetaEditorPanel = useMemo(() => {
-    const hasManualThumbnail = postThumbnailUrl.trim().length > 0
-    const hasFirstBodyImage = deferredContentDerived.firstImage.trim().length > 0
-    const sourceTone = hasManualThumbnail ? "manual" : hasFirstBodyImage ? "auto" : "empty"
-    const sourceLabel = hasManualThumbnail
-      ? "수동 썸네일 URL 사용 중"
-      : hasFirstBodyImage
-        ? "본문 첫 이미지를 자동으로 사용 중"
-        : "현재 썸네일이 없습니다"
-
-    return (
-      <PreviewEditorSection onPasteCapture={handleThumbnailPaste}>
-        <PreviewEditorSectionHeader>
-          <strong>썸네일 이미지</strong>
-          <ThumbnailSourceStatus data-tone={sourceTone}>{sourceLabel}</ThumbnailSourceStatus>
-        </PreviewEditorSectionHeader>
-        <FieldLabel htmlFor="post-thumbnail-url-modal">썸네일 URL</FieldLabel>
-        <Input
-          id="post-thumbnail-url-modal"
-          placeholder="https://... (비우면 본문 첫 이미지 자동 사용)"
-          value={postThumbnailUrl}
-          onChange={(e) => handleThumbnailUrlModalChange(e.target.value)}
-          onPaste={handleThumbnailPaste}
-        />
-        <MetaPrimaryActionRow>
-          <PrimaryButton
-            type="button"
-            title={POST_IMAGE_UPLOAD_RULE_LABEL}
-            disabled={isThumbnailUploadDisabled}
-            onClick={openThumbnailFileInput}
-          >
-            {loadingKey === "uploadThumbnail" ? "업로드 중..." : "썸네일 파일 업로드"}
-          </PrimaryButton>
-        </MetaPrimaryActionRow>
-        <MetaSecondaryActionRow>
-          <Button type="button" disabled={!hasFirstBodyImage} onClick={applyFirstBodyImageToThumbnail}>
-            본문 첫 이미지 사용
-          </Button>
-          <Button type="button" data-variant="text" onClick={resetThumbnailToAutoMode}>
-            자동 모드로 되돌리기
-          </Button>
-        </MetaSecondaryActionRow>
-        {thumbnailImageFileName ? <FieldHelp>선택 파일: {thumbnailImageFileName}</FieldHelp> : null}
-      </PreviewEditorSection>
-    )
-  }, [
-    applyFirstBodyImageToThumbnail,
-    deferredContentDerived.firstImage,
-    handleThumbnailUrlModalChange,
-    isThumbnailUploadDisabled,
-    loadingKey,
-    openThumbnailFileInput,
-    handleThumbnailPaste,
-    postThumbnailUrl,
-    resetThumbnailToAutoMode,
-    thumbnailImageFileName,
-  ])
+  const thumbnailEditorPanel = (
+    <EditorStudioThumbnailEditorPanel
+      finalizePreviewThumbPointer={finalizePreviewThumbPointer}
+      handlePreviewThumbPointerDown={handlePreviewThumbPointerDown}
+      handlePreviewThumbPointerMove={handlePreviewThumbPointerMove}
+      isPreviewThumbDragging={isPreviewThumbDragging}
+      isPreviewThumbnailError={isPreviewThumbnailError}
+      postThumbnailZoom={postThumbnailZoom}
+      previewThumbFrameRef={previewThumbFrameRef}
+      safePreviewThumbnail={safePreviewThumbnail}
+      setIsPreviewThumbnailError={setIsPreviewThumbnailError}
+      onThumbnailZoomChange={handleThumbnailZoomModalChange}
+      onResetThumbnailZoom={resetThumbnailZoomInModal}
+    />
+  )
+  const previewMetaEditorPanel = (
+    <EditorStudioThumbnailMetaPanel
+      firstBodyImageUrl={deferredContentDerived.firstImage}
+      isThumbnailUploadDisabled={isThumbnailUploadDisabled}
+      isThumbnailUploading={loadingKey === "uploadThumbnail"}
+      postThumbnailUrl={postThumbnailUrl}
+      thumbnailImageFileName={thumbnailImageFileName}
+      thumbnailUploadRuleLabel={POST_IMAGE_UPLOAD_RULE_LABEL}
+      onApplyFirstBodyImage={applyFirstBodyImageToThumbnail}
+      onOpenThumbnailFileInput={openThumbnailFileInput}
+      onResetThumbnailToAutoMode={resetThumbnailToAutoMode}
+      onThumbnailPaste={handleThumbnailPaste}
+      onThumbnailUrlChange={handleThumbnailUrlModalChange}
+    />
+  )
   const editorPrimaryActionType: PublishActionType =
     editorMode === "create" ? "create" : isTempDraftMode ? "temp" : "modify"
   const editorPrimaryActionLabel =
@@ -6157,107 +6065,6 @@ const CompactPublishEditorToggle = styled.button`
   }
 `
 
-const PreviewEditorSection = styled.div`
-  display: grid;
-  gap: 0.58rem;
-  min-width: 0;
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  border-radius: 12px;
-  background: ${({ theme }) => theme.colors.gray2};
-  padding: 0.85rem;
-`
-
-const PreviewEditorSectionHeader = styled.div`
-  display: grid;
-  gap: 0.14rem;
-
-  strong {
-    color: ${({ theme }) => theme.colors.gray12};
-    font-size: 0.86rem;
-    font-weight: 700;
-    line-height: 1.3;
-  }
-
-  span {
-    color: ${({ theme }) => theme.colors.gray11};
-    font-size: 0.74rem;
-    line-height: 1.45;
-  }
-`
-
-const PreviewThumbFrame = styled.div`
-  --preview-thumb-width: 100%;
-  --preview-thumb-height: 100%;
-  --preview-thumb-left: 0%;
-  --preview-thumb-top: 0%;
-
-  position: relative;
-  width: min(100%, 360px);
-  justify-self: start;
-  aspect-ratio: 1.94 / 1;
-  border-radius: ${({ theme }) => `${theme.variables.ui.card.radius}px`};
-  border: ${({ theme }) => `${theme.variables.ui.card.borderWidth}px solid ${theme.colors.gray4}`};
-  background: ${({ theme }) => theme.colors.gray4};
-  overflow: hidden;
-  user-select: none;
-  isolation: isolate;
-
-  &[data-draggable="true"] {
-    cursor: grab;
-    touch-action: none;
-  }
-
-  &[data-dragging="true"] {
-    cursor: grabbing;
-  }
-
-  @media (max-width: 780px) {
-    width: 100%;
-  }
-
-  img {
-    position: absolute;
-    display: block;
-    pointer-events: none;
-    user-select: none;
-    touch-action: none;
-    -webkit-user-drag: none;
-    will-change: top, left, width, height;
-  }
-
-  &::after {
-    content: "";
-    position: absolute;
-    inset: 0;
-    background: linear-gradient(180deg, rgba(0, 0, 0, 0) 45%, rgba(0, 0, 0, 0.16) 100%);
-    opacity: 0.9;
-    pointer-events: none;
-  }
-
-  .placeholder {
-    width: 100%;
-    height: 100%;
-    display: grid;
-    place-content: center;
-    text-align: center;
-    gap: 0.24rem;
-    padding: 0.7rem;
-
-    em {
-      font-style: normal;
-      color: ${({ theme }) => theme.colors.gray10};
-      font-weight: 700;
-      font-size: 0.84rem;
-    }
-
-    span {
-      color: ${({ theme }) => theme.colors.gray11};
-      font-size: 0.74rem;
-      line-height: 1.4;
-    }
-  }
-`
-
 const SummaryCounter = styled.span`
   justify-self: end;
   color: ${({ theme }) => theme.colors.gray10};
@@ -6316,99 +6123,6 @@ const ComposeStatusRow = styled.div`
     span {
       color: ${({ theme }) => theme.colors.red11};
     }
-  }
-`
-
-const ZoomControlRow = styled.div`
-  display: grid;
-  gap: 0.42rem;
-  align-items: start;
-  justify-items: start;
-  width: min(100%, 360px);
-
-  @media (max-width: 780px) {
-    width: 100%;
-  }
-`
-
-const ZoomRangeInput = styled.input`
-  width: 100%;
-  accent-color: ${({ theme }) => theme.colors.blue9};
-`
-
-const ZoomValue = styled.span`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 30px;
-  border-radius: 999px;
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  padding: 0 0.55rem;
-  color: ${({ theme }) => theme.colors.gray11};
-  font-size: 0.78rem;
-  line-height: 1;
-  font-weight: 700;
-`
-
-const ZoomControlMeta = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0.55rem;
-  flex-wrap: wrap;
-`
-
-const ThumbnailSourceStatus = styled.span`
-  display: inline-flex;
-  align-items: center;
-  width: fit-content;
-  min-height: 24px;
-  border-radius: 999px;
-  border: 1px solid ${({ theme }) => theme.colors.gray6};
-  padding: 0 0.54rem;
-  color: ${({ theme }) => theme.colors.gray10};
-  font-size: 0.72rem;
-  font-weight: 700;
-  line-height: 1;
-
-  &[data-tone="manual"] {
-    border-color: ${({ theme }) => theme.colors.blue7};
-    background: ${({ theme }) => theme.colors.blue3};
-    color: ${({ theme }) => theme.colors.blue11};
-  }
-
-  &[data-tone="auto"] {
-    border-color: ${({ theme }) => theme.colors.green7};
-    background: ${({ theme }) => theme.colors.green3};
-    color: ${({ theme }) => theme.colors.green11};
-  }
-`
-
-const MetaPrimaryActionRow = styled.div`
-  display: grid;
-
-  > button {
-    width: 100%;
-    justify-content: center;
-    text-align: center;
-    white-space: normal;
-    line-height: 1.32;
-  }
-`
-
-const MetaSecondaryActionRow = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.55rem;
-
-  > button:first-of-type {
-    flex: 1 1 11rem;
-  }
-
-  > button[data-variant="text"] {
-    flex: 0 0 auto;
-    text-align: left;
   }
 `
